python_service/requirements.txt
fastapi
supabase
openai
python-dotenv
tqdm
uvicorn
requests
beautifulsoup4
python-dotenv
websockets